feat: add CalendarContext and extract Calendar subcomponents
